Adds a pokemob cache for restoring lost pokemobs.
Pokemobs are added/updated to this cache whenever they pop from
pokecube.  This only caches player owned pokemobs, and the command to
restore has 2 options:

check which clicking on the pokemob name teleports you to the last
popped location for the pokemob, and give, which gives you the pokecube
of the pokemob.